requirements.txt
crewai
crewai-tools
google-adk
google-generativeai
langchain
langchain-openai
langchain-community
duckduckgo-search
langgraph
wikipedia